feat(Dataworker#index): Add buffer getChallengeRemaining calls (#2426)
* Revert "feat(TransactionUtils): Do not send `error` level log in `runTransaction()` (#2416)"

This reverts commit 4a9fe99c9dc96e9400ad300ed46d920044b0a84b.

* feat(Dataworker): Add Mutex for L1-executor to try to reduce collisions

Use a simple mutex to try to reduce L1 executor collisions. Default settings are to lock for 5 mins. Mutex can be disabled with environment variable.

* Update index.ts

* Update index.ts

* feat(Dataworker#index): Add buffer getChallengeRemaining calls

The current bot seems to attempt to submit a transaction before the challenge period has actually passed, therefore I suspect that there is some mysterious error happening in the `getChallengeRemaining()`'s parallel calls to `getCurrentTime` and `pendingRootBundleProposal`. The challenge remaining returns as 0 but when the multicaller client goes to execute the transactions, empirically we're seeing about once a day a failure related to the challenge period liveness not being over on-chain.

This PR adds more robust logs as well as a buffer of 2 blocks.

* Update index.ts

* Revert "Update index.ts"

This reverts commit 222b527a4cb8e3283f954dc552ba3d6fa2ad78bc.

* Update index.ts

* Update index.ts

* Add explicit block tag to help with logs