improve(relayer): Skip benevolent repayment for fast routes (#2496)
Avoid taking repayment on 7 day withdrawal chains where a fast rebalance
route is available for the given token (i.e. USDC on CCTP chains, USDT
on OFT chains). This increases the chance of taking repayment on a chain
where it would actually make a difference for HubPool utilisation.

Additionally, update the list of slow withdrawal chains to permit the 
relayer to hold greater balances on chains where it makes a difference
for HubPool utilisation.