feat(InventoryClient): Simplify approximation of upcoming refunds
Removes dependency on BundleDataClient to approximate upcoming refunds in order to better select repayment chain choice based on virtual balances.

Currently we approximate refunds by (1) querying Arweave to get upcoming refund data for the pending and last executed bundles and this value to (2) an approximation of all upcoming fills that we have sent since the last pending bundle was proposed.

This PR simplifies this algorithm considerably by:
- For each chain that we care about
- Find the last RelayedRootBundle event to that l2 chain
- Assume all fills sent by `this.relayer` since that last-relayed-bundle's end block are valid and assume they will result in upcoming refunds.