Add comments and protection about usage of `getL1TokenAddress`
This function is dangerous and can throw so we need to make sure we are catching failure possibilities before hand. We already do so today so this mostly adds comments.

To this specific PR, I mistakenly assumed that it was safe because the first use case, from `determineRefundChain()` is safe because it previously calls `depositForcesOriginChainRepayment`, but `_getLatestRunningBalances()` might not filter it correctly.