improve(relayer): Don't auto-bail on block error (#2549)
The existing hard-stop implementation might be too strict and causing
fragility in the relayer. viem is OK to reconnect in the background and
continue listening for new blocks. This doesn't affect re-org detection
since that's based on the individual events; in fact it allows more
time for a re-org to be detected.

This change is prompted by an odd pairing of zkSync & QuickNode, which
tends to cause a lot of JSON parsing errors to be thrown. Unfortunately
it's unclear exactly which response is causing the error, investigation
ongoing.