fix(zkstack-finalizer): Add special case for USDC withdrawals
USDC withdrawals require special handling because the L2 USDC bridge doesn't have an `l1SharedBridge` method but instead has a `l1USDCBridge` method. The rest of the process is the same.

This PR keeps the `zksync` finalizer adapter very light weight but we should replace it eventually with a bump to `zksync-ethers>=6.20.0`