improve(monitor): Permit blanket opt-out of follow distance (#2275)
The follow distance can cause under-reporting on pending repayments for
any fills that were made within the default follow distance. In general
this is undesirable for the monitor. The operator could override this
manually for each chain, but that's a maintenance overhead when new
chains are added. Instead, permit a global opt-out of the follow
distance configurable.