improve(monitor): Use BundleDataApproxClient to estimate upcoming refunds (#2553)
* feat(InventoryClient): Simplify approximation of upcoming refunds

Removes dependency on BundleDataClient to approximate upcoming refunds in order to better select repayment chain choice based on virtual balances.

Currently we approximate refunds by (1) querying Arweave to get upcoming refund data for the pending and last executed bundles and this value to (2) an approximation of all upcoming fills that we have sent since the last pending bundle was proposed.

This PR simplifies this algorithm considerably by:
- For each chain that we care about
- Find the last RelayedRootBundle event to that l2 chain
- Assume all fills sent by `this.relayer` since that last-relayed-bundle's end block are valid and assume they will result in upcoming refunds.

* Fix tests

* fix

* Update src/clients/InventoryClient.ts

Co-authored-by: Paul <108695806+pxrl@users.noreply.github.com>

* Add comments and protection about usage of `getL1TokenAddress`

This function is dangerous and can throw so we need to make sure we are catching failure possibilities before hand. We already do so today so this mostly adds comments.

To this specific PR, I mistakenly assumed that it was safe because the first use case, from `determineRefundChain()` is safe because it previously calls `depositForcesOriginChainRepayment`, but `_getLatestRunningBalances()` might not filter it correctly.

* Update src/clients/InventoryClient.ts

Co-authored-by: Paul <108695806+pxrl@users.noreply.github.com>

* Update InventoryClient.ts

* Add unit tests

* Fix filters

* Update InventoryClient.ts

* Fix

* fix

* Update InventoryClient.ts

* Update InventoryClient.UpcomingRefunds.ts

* Filter for refunds for all relayers for computing running balances

* Catch `getL1TokenAddress` if it fails

* Update src/clients/InventoryClient.ts

Co-authored-by: Paul <108695806+pxrl@users.noreply.github.com>

* Update src/clients/InventoryClient.ts

Co-authored-by: Paul <108695806+pxrl@users.noreply.github.com>

* Update InventoryClient.ts

* Handle getL1TokenAddress() returning undefined

* Refactor into new client

* Update InventoryClient.ts

* Make functions private

* fix tests

* Handle edge cases for tests

* improve(monitor): Use BundleDataApproxClient to estimate upcoming refunds

The motivation for this is to remove dependence on the BundleDataClient in the Monitor. With this PR we can delete most of the helper functions in the BundleDataClient that are only used by the Monitor

* Update InventoryClient.ts

* fix types

* Update Monitor.ts

* Update BundleDataApproxClient.ts

* Use getBundleEndBlockForChain

* Update BundleDataApproxClient.ts

* Fix test

* Fix monitor test

* Update validateRunningBalances.ts

* Update validateRunningBalances.ts

* Update InventoryClient.ts

* fix monitor

* Update src/monitor/Monitor.ts

Co-authored-by: Paul <108695806+pxrl@users.noreply.github.com>

* add init()

* Revert "add init()"

This reverts commit c560f8c13bec5084829e01d530c652d63c2583af.

* Update Monitor.ts

* Index relayers in report by native address

* Update Monitor.ts

* Update Monitor.ts

* Update Monitor.ts

---------

Co-authored-by: Paul <108695806+pxrl@users.noreply.github.com>