fix(rebalancer): Don't flush TokenClient state prematurely (#2606)
Flushing TokenClient state before considering withdrawals means that the 
rebalancer assumes that there are no deposits awaiting to be filled.
This can cause the rebalancer to pull back tokens that were intended to
be used by the relayer to fill shortfalls. The rebalancer should instead
factor in any simulated fills that were made as part of identifying
shortfalls.