improve(ProfitClient): Better inference of native gas tokens
This area has been an issue for custom gas tokens on Solana, HyperEVM
and Plasma. Always try to grab the mainnet address for the native gas
token, and revert to the token address on the destination chain if that
fails. This relies on the wrapped and native token addresses having the
same configurations in the constants repo; for now that holds true.