improve(TransactionUtils): Overhaul error handling on transaction submission (#2126)
Implement more fine-grained handling based of the errors that can be
emitted when simulating or submitting transactions.

This permits more intelligent behaviour on REPLACEMENT_UNDERPRICED and
NONCE_EXPIRED errors, as well as recursive pricing updates if the
transaction price for a non-replacement is insufficient.

Ideally this entirely eliminates the phenomen of the confirmed & pending
nonces getting completely out of sync, and in cases where they do, the
bots should recover automatically. It should also permit the various
chain priority fee scalers to be reset to 1 (or at least very close to
1). Additionally, the improved logging should make it easier to diagnose
residual unusual/disruptive behaviour.