feat: Rebalancer should rebalance for shortfalls separately from inventory requirements
This is a subtle change but a big improvement to capital efficiency. Currently our heuristic for deciding to rebalance is when an L2's inventory minus any pending deposit amounts to that chain is below some target %. We then rebalance the total deficit %, however, sometimes this deficit amount can be very large especially when the chain's balance is already underallocated when not even accounting for the short fall.

We've run into situations in the past where we've refused to rebalance even if we have enough to cover a shortfall but not enough to also cover for the inventory underallocation.

This PR changes the behavior to first evaluate rebalances for shortfalls individually, and then cover inventory rebalances.