feat: Create Refiller bot (#2593)
* feat(Monitor): Incorporate SwapHelper into Monitor.refiller

Eventually we can move the refiller logic out of the Monitor but I wanted to test this PoC out in the Monitor without adding too much code. Assuming the SwapClient works well we can productionize it into a separate bot in a follow up PR

* Update Monitor.ts

* Fixes

* Update Monitor.ts

* Update Monitor.ts

* Update Monitor.ts

* fix

* feat: Create Refiller bot

This removes all the `refillBalances` logic from the Monitor and moves it into its own bot. This should be much easier to maintain

* Remove monitor stuff

* Add simple test

* Simplify

* Update Refiller.ts

* Update index.ts

* Update Refiller.ts

* Add comments about flipping if-else blocks to remove indentation