feat: Rebalancer should rebalance for shortfalls separately from inventory requirements (#2626)
* feat: Add any->any finalizer strategy

This is prerequisite work to supporting any -> any rebalancing, we first should have a working finalizer.

This PR adds an `any->any` finalization mode. I chose to go with this implementation because eventually we can replace the `l1->l2` and `l2->l1` strategies with this new mode.

* feat: Support CCTP V2 fast transfers

* Add L1 to L2

* fix

* fix

* reset finalizer

* Delete anyToAny.ts

* Update CCTPUtils.ts

* Fix tests

* Fix tests and pending transfer tracking

* Update UsdcCCTPBridge.ts

* Update MockCctpEvents.sol

* Update AdapterManager.SendTokensCrossChain.ts

* Update MockCctpEvents.sol

* WIP

* Update UsdcCCTPBridge.ts

* fix

* fix

* feat: Rebalancer should rebalance for shortfalls separately from inventory requirements

This is a subtle change but a big improvement to capital efficiency. Currently our heuristic for deciding to rebalance is when an L2's inventory minus any pending deposit amounts to that chain is below some target %. We then rebalance the total deficit %, however, sometimes this deficit amount can be very large especially when the chain's balance is already underallocated when not even accounting for the short fall.

We've run into situations in the past where we've refused to rebalance even if we have enough to cover a shortfall but not enough to also cover for the inventory underallocation.

This PR changes the behavior to first evaluate rebalances for shortfalls individually, and then cover inventory rebalances.

* Update InventoryClient.InventoryRebalance.ts

* fix

* Update InventoryClient.ts

* Refactor

* Update InventoryClient.ts

* Update InventoryClient.ts

* reduce log

* Update InventoryClient.ts

* Update InventoryClient.ts

* Update TokenClient.ts

* fix

* Update InventoryClient.ts

* Update InventoryClient.InventoryRebalance.ts

* Update InventoryClient.ts