improve(Finalizer): Remove redundant individual transaction simulation step
The MulticallerClient already simulates transactions and in all cases, we submit transactions via this client. Remove the outdated TransactionClient logic.

This also removes the check on the total gas cost of the aggregated batch finalizations, but this should be less necessary as we run the finalizer more frequently and have fewer large batches than before.

One alternative is to have a hardcoded cap onthe # of finalizations to send at once, this could be useful.