improve(finalizer): Autogenerate finalizer configs
Having to update the finalizer config manually when adding new chains
consumes time and can be forgotten, leading to subsequent issues
post-deployment. A lot of the configuration for a new chain can be
automated based on its chain family and CCTP domain ID configuration,
specified upstream in the constants repo. Lean on this in order to
reduce the need for manual updates here.

For follow-up:
- Handle inconsistencies like Arbitrum, Polygon & zkSync.