improve(ProfitClient): Trim unprofitable fill logs (#2664)
Don't dump the entire message into a log message because it causes
truncation. Instead, just log the messageHash; that can be used to infer
whether a message is present or not.

Also, it's OK to drop conversion to bytes32 here. Logging works OK
without it. 

This change motivated by a perceived increase in the number of deposits
w/ messages that are reverting in simulation.