improve(constants): Autogenerate RPC getLogs ranges (#2656)
Most of these configs are set to 10k and are typically overridden in the
environment if a particular RPC provider imposes a smaller limit. It's
easier to deal with the exceptions and otherwise assume a limit of 10k.

HyperEVM sticks out as an annoying constraint of 1k blocks, imposed by
QuickNode.