refactor(relayer): Split out block & event subscription (#2686)
One step towards enabling the parent process to decide dynamically which
subscriptions should be made.

While working on this it was difficult to factor out common parts of the
websocket and https-based listener implementations. In the end it became
apparent that the files were almost identical, so instead they have been
unified with a runtime check on the configured transport (default wss,
can be overridden to https). This addresses some longstanding tech debt.