feat: Add EventListener class
This is a simple class that can be instantiated to get event
notification on new blocks or new events. Implementation is
predominantly based on the existing relayer listener, with some minor
differences. Refactoring will be done to integrate this into the relayer
listener implementation so as to simplify it.

There's also a bit to follow-up on, including:
- Handling event unsubscriptions.
- Handling errors.

See scripts/eventListener.ts for an example of how to use this.