improve(relayer): Catch event scraping exceptions (#2721)
Relayer listener processes can sometimes throw on RPC issues. In most
cases this seems to occur due to quorum disagreemnents during the
initial scraping of historical events. It's an improvement to instead
just return no events for the failed scenario, and proceed to listening
for new events. This should improve resilience & fault tolerance in the
relayer.