feat: Allow user to set specific output tokens to fill
Can be used to more easily support token swaps. We currently use RELAYER_TOKENS to control which input tokens to fill for and assume that every output token is equivalent to the input token.

Therefore, in order to support swaps we need to also add a filter for output tokens