feat(relayer): Event-based HubPoolClient/ConfigStoreClient updates
The relayer currently polls on each loop for new ConfigStoreClient
updates. In the event that a new Hub chain block is detected, it then
proceeds to update the HubPoolClient. On a new block, this can take 1 -
2 seconds. This directly impacts the relayer's fill performance and can
cause delays.

This change makes updates event-driven, but hooking into new block
events emitted by the mainnet SpokePoolClient. ConfigStoreClient and
HubPoolClient updates are then performed in the background.

There are some minor concurrency issues to be aware of here, though
nothing that's expected to cause problems in production. This feature is
however guarded behind config so that it's opt-in and can be disabled on
short notice.