improve(Relayer): Optimise TokenClient updates (#2767)
Restrict updates to configured destination chains (or all chains, if not
specified). There's no point updating token balances on chains where the
relayer is not actively filling. This gives a measurable speed
improvement on a task that is repeated frequently.

The real enemy here is that the tokenClient updates on the hot path; it
should update in the background on a separate schedule, but there are
some concurrency issues to be checked first.