chore: Revert "Bump beta contracts (#2817)"
This reverts commit 818d18e7e4e2c6282266cc0dffc2cf35de74a59d.

Also update the dstCctpHandler address with the latest deployed version.
This can't yet be imported from a published contracts release due to
typechain issues. While here, squash dupliate address specifications for
the same contract. The naming misalignment is annoying, but it's
probably better to handle it in code rather than duplicate configs.