improve(hlFinalizer): Check for finalisation on each nth block (#2819)
Update the hyperliquid finalizer to use a similar persistent-handover
pattern to what is implemented in the relayer. This PR attempts to
improve on what's used in the relayer by factoring out some of the
complexity to a separate class. It's a mixed success and definitely
needs more iteration, but in local testing it's working OK, despite the
warts.